refactor: (WIP) remove global error handler and implement auth interceptor; update auth state management
