+++ src/app/auth/auth.guard.ts
@@ -1,6 +1,6 @@
 import { Injectable, inject } from '@angular/core';
 import { CanActivateFn, Router } from '@angular/router';
-import { AuthService } from './auth.service';
+import { AuthService } from './services/auth.service';
 import { AuthStatus } from './enums/auth-status.enum';
 
 export const authGuard: CanActivateFn = (route, state) => {
