+++ src/app/shared/interceptors/auth.interceptor.ts
@@ -3,17 +3,15 @@ import {
   HttpHandler,
   HttpInterceptor,
   HttpRequest,
-  HttpErrorResponse
 } from '@angular/common/http';
-import { Injectable, Injector } from '@angular/core';
-import { Observable, throwError } from 'rxjs';
-import { catchError, switchMap } from 'rxjs/operators';
-import { AuthService } from '../../auth/auth.service';
+import { inject, Injectable } from '@angular/core';
+import { Observable } from 'rxjs';
+import { AuthService } from '../../auth/services/auth.service';
 
 @Injectable()
 export class AuthInterceptor implements HttpInterceptor {
   private readonly _EXCLUDED_BEARER_ROUTES = ['/auth'];
-  constructor(private injector: Injector) {}
+  private readonly _authService = inject(AuthService);
 
   intercept(
     req: HttpRequest<any>,
@@ -21,46 +19,17 @@ export class AuthInterceptor implements HttpInterceptor {
   ): Observable<HttpEvent<any>> {
     if (this.shouldExcludeBearerRoute(req.url)) return next.handle(req);
 
-    const authService = this.injector.get(AuthService);
-    const accessToken = authService.accessToken();
+    const accessToken = this._authService.accessToken();
     if (!accessToken) return next.handle(req);
 
     const authReq = req.clone({
       setHeaders: { Authorization: `Bearer ${accessToken}` },
     });
 
-    return next.handle(authReq).pipe(
-      catchError((error: HttpErrorResponse) => {
-        if (error.status === 401) {
-          return authService.refreshTokenRequest().pipe(
-            switchMap(() => {
-              const newAccessToken = authService.accessToken();
-              const retryReq = req.clone({
-                setHeaders: { Authorization: `Bearer ${newAccessToken}` },
-              });
-              return next.handle(retryReq);
-            }),
-            catchError((refreshError) => {
-              authService.logout();
-              return throwError(() => refreshError);
-            })
-          );
-        }
-        return throwError(() => error);
-      })
-    );
+    return next.handle(authReq);
   }
 
   private shouldExcludeBearerRoute(url: string): boolean {
-    try {
-      const path = new URL(url, window.location.origin).pathname;
-      return this._EXCLUDED_BEARER_ROUTES.some((route) =>
-        path.startsWith(route)
-      );
-    } catch {
-      return this._EXCLUDED_BEARER_ROUTES.some((route) =>
-        url.startsWith(route)
-      );
-    }
+    return this._EXCLUDED_BEARER_ROUTES.some((route) => url.startsWith(route));
   }
 }
